Loading Data

Load all input files.

Data_loading(
  Candidate_Gene_file_path,
  Candidate_Peak_file_path,
  scRNA_readcount_file_path,
  scRNA_Gene_file_path,
  scRNA_cellmeta_file_path,
  scATAC_readcount_file_path,
  scATAC_Peak_file_path,
  scATAC_cellmeta_file_path,
  Motif_mapping_file_path,
  Motif_name_file_path,
  Ref_seq_file_path
)

Arguments

scRNA_readcount_file_path, scRNA_gene_file_path, scRNA_cellmeta_file_path

paths to filtered scRNA data of the cell type

scATAC_readcount_file_path, scATAC_peak_file_path, scATAC_cellmeta_file_path

paths to filtered scATAC data of the cell type

Motif_mapping_file_path, Motif_name_file_path

paths to TF motif prior on all ATAC peaks

Ref_seq_file_path

path to the Refseq file for transcription starting site extraction

Candidate_gene_file_path, Candidate_peak_file_path

paths to pre-selected candidate genes and peaks for the cell type

Value

A list named "loaded_data" containing the following elements:

Common_samples: a vector of common samples between scRNA and scATAC data

Candidate_Genes: a data frame of candidate genes

Candidate_Peaks: a data frame of candidate peaks

scRNA_Genes: a data frame of scRNA genes

scRNA_cells: a data frame of scRNA cells

scRNA_read_count_matrix: a sparse matrix of scRNA read counts

scATAC_Peaks: a data frame of scATAC peaks

scATAC_cells: a data frame of scATAC cells

scATAC_read_count_matrix: a sparse matrix of scATAC read counts

Motifs: a data frame of TF motifs

TF_peak_binding_matrix: a sparse matrix of TF binding on ATAC peaks

Refseq: a data frame of Refseq